+++ back-end/src/api/app.js
@@ -1,6 +1,10 @@
 const express = require('express');
 const cors = require('cors');
-const { registerRoute, loginRoute, custumerRoute, checkoutRoute } = require('../routes/index');
+const { registerRoute,
+    loginRoute,
+    custumerRoute,
+    salesRoute,
+    salesProductsRoute } = require('../routes/index');
 const { uniqueConstraintError } = require('../middlewares/errorMiddleware');
 
 const app = express();
@@ -10,7 +14,8 @@ app.use(cors());
 app.use('/register', registerRoute);
 app.use('/login', loginRoute);
 app.use('/custumer', custumerRoute);
-app.use('/checkout', checkoutRoute);
+app.use('/sales', salesRoute);
+app.use('/salesProducts', salesProductsRoute);
 
 app.use(uniqueConstraintError);
