+++ back-end/src/controllers/salesController.js
@@ -1,6 +1,6 @@
 const { StatusCodes } = require('http-status-codes');
 const rescue = require('express-rescue');
-const service = require('../services/checkoutService');
+const service = require('../services/salesService');
 
 const createSale = rescue(async (req, res) => {
     const payload = req.body;
@@ -8,13 +8,6 @@ const createSale = rescue(async (req, res) => {
     res.status(StatusCodes.CREATED).json(result);
 });
 
-const createSalesProducts = rescue(async (req, res) => {
-    const payload = req.body;
-    const result = await service.createSalesProducts(payload);
-    res.status(StatusCodes.CREATED).json(result);
-});
-
 module.exports = {
     createSale,
-    createSalesProducts,
 };
