+++ back-end/src/routes/index.js
@@ -1,6 +1,7 @@
 const registerRoute = require('./registerRoute');
 const loginRoute = require('./loginRoute');
 const custumerRoute = require('./custumerRoute');
-const checkoutRoute = require('./checkoutRoute');
+const salesRoute = require('./salesRoute');
+const salesProductsRoute = require('./salesProductsRoute');
 
-module.exports = { registerRoute, loginRoute, custumerRoute, checkoutRoute };
+module.exports = { registerRoute, loginRoute, custumerRoute, salesRoute, salesProductsRoute };
