+++ back-end/src/routes/salesProductsRoute.js
@@ -2,9 +2,8 @@ const express = require('express');
 
 const router = express.Router();
 
-const controller = require('../controllers/checkoutController');
+const controller = require('../controllers/salesProductsController');
 
-router.post('/', controller.createSale);
 router.post('/', controller.createSalesProducts);
 
 module.exports = router;
