--- back-end/src/services/checkoutService.js
@@ -1,10 +1,0 @@
-const { sale, salesProduct } = require('../database/models');
-
-const createSale = async (payload) => sale.create(payload);
-
-const createSalesProducts = async (payload) => salesProduct.create(payload);
-
-module.exports = {
-    createSale,
-    createSalesProducts,
-};
