+++ back-end/src/services/salesProductsService.js
@@ -1,0 +1,7 @@
+const { salesProduct } = require('../database/models');
+
+const createSalesProducts = async (payload) => salesProduct.create(payload);
+
+module.exports = {
+    createSalesProducts,
+};
