+++ back-end/src/services/salesService.js
@@ -1,0 +1,7 @@
+const { sale } = require('../database/models');
+
+const createSale = async (payload) => sale.create(payload);
+
+module.exports = {
+    createSale,
+};
